LayoutTests/platform/chromium/svg/custom/global-constructors-expected.txt

Test to make sure we expose all the global constructor objects requested by http://www.w3.org/TR/SVG11/ecmascript-binding.html

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL SVGException.toString() should be [object SVGExceptionConstructor]. Was function SVGException() { [native code] }.
FAIL SVGElement.toString() should be [object SVGElementConstructor]. Was function SVGElement() { [native code] }.
FAIL SVGAnimatedBoolean.toString() should be [object SVGAnimatedBooleanConstructor]. Was function SVGAnimatedBoolean() { [native code] }.
FAIL SVGAnimatedString.toString() should be [object SVGAnimatedStringConstructor]. Was function SVGAnimatedString() { [native code] }.
FAIL SVGStringList.toString() should be [object SVGStringListConstructor]. Was function SVGStringList() { [native code] }.
FAIL SVGAnimatedEnumeration.toString() should be [object SVGAnimatedEnumerationConstructor]. Was function SVGAnimatedEnumeration() { [native code] }.
FAIL SVGAnimatedInteger.toString() should be [object SVGAnimatedIntegerConstructor]. Was function SVGAnimatedInteger() { [native code] }.
FAIL SVGNumber.toString() should be [object SVGNumberConstructor]. Was function SVGNumber() { [native code] }.
FAIL SVGAnimatedNumber.toString() should be [object SVGAnimatedNumberConstructor]. Was function SVGAnimatedNumber() { [native code] }.
FAIL SVGNumberList.toString() should be [object SVGNumberListConstructor]. Was function SVGNumberList() { [native code] }.
FAIL SVGAnimatedNumberList.toString() should be [object SVGAnimatedNumberListConstructor]. Was function SVGAnimatedNumberList() { [native code] }.
FAIL SVGLength.toString() should be [object SVGLengthConstructor]. Was function SVGLength() { [native code] }.
FAIL SVGAnimatedLength.toString() should be [object SVGAnimatedLengthConstructor]. Was function SVGAnimatedLength() { [native code] }.
FAIL SVGLengthList.toString() should be [object SVGLengthListConstructor]. Was function SVGLengthList() { [native code] }.
FAIL SVGAnimatedLengthList.toString() should be [object SVGAnimatedLengthListConstructor]. Was function SVGAnimatedLengthList() { [native code] }.
FAIL SVGAngle.toString() should be [object SVGAngleConstructor]. Was function SVGAngle() { [native code] }.
FAIL SVGAnimatedAngle.toString() should be [object SVGAnimatedAngleConstructor]. Was function SVGAnimatedAngle() { [native code] }.
FAIL SVGColor.toString() should be [object SVGColorConstructor]. Was function SVGColor() { [native code] }.
FAIL SVGICCColor.toString() should be [object SVGICCColorConstructor]. Threw exception ReferenceError: SVGICCColor is not defined
FAIL SVGRect.toString() should be [object SVGRectConstructor]. Was function SVGRect() { [native code] }.
FAIL SVGAnimatedRect.toString() should be [object SVGAnimatedRectConstructor]. Was function SVGAnimatedRect() { [native code] }.
FAIL SVGStylable.toString() should be [object SVGStylableConstructor]. Threw exception ReferenceError: SVGStylable is not defined
FAIL SVGLocatable.toString() should be [object SVGLocatableConstructor]. Threw exception ReferenceError: SVGLocatable is not defined
FAIL SVGTransformable.toString() should be [object SVGTransformableConstructor]. Threw exception ReferenceError: SVGTransformable is not defined
FAIL SVGTests.toString() should be [object SVGTestsConstructor]. Threw exception ReferenceError: SVGTests is not defined
FAIL SVGLangSpace.toString() should be [object SVGLangSpaceConstructor]. Threw exception ReferenceError: SVGLangSpace is not defined
FAIL SVGExternalResourcesRequired.toString() should be [object SVGExternalResourcesRequiredConstructor]. Threw exception ReferenceError: SVGExternalResourcesRequired is not defined
FAIL SVGFitToViewBox.toString() should be [object SVGFitToViewBoxConstructor]. Threw exception ReferenceError: SVGFitToViewBox is not defined
FAIL SVGZoomAndPan.toString() should be [object SVGZoomAndPanConstructor]. Was function SVGZoomAndPan() { [native code] }.
FAIL SVGViewSpec.toString() should be [object SVGViewSpecConstructor]. Was function SVGViewSpec() { [native code] }.
FAIL SVGURIReference.toString() should be [object SVGURIReferenceConstructor]. Threw exception ReferenceError: SVGURIReference is not defined
FAIL SVGCSSRule.toString() should be [object SVGCSSRuleConstructor]. Threw exception ReferenceError: SVGCSSRule is not defined
FAIL SVGDocument.toString() should be [object SVGDocumentConstructor]. Was function SVGDocument() { [native code] }.
FAIL SVGSVGElement.toString() should be [object SVGSVGElementConstructor]. Was function SVGSVGElement() { [native code] }.
FAIL SVGGElement.toString() should be [object SVGGElementConstructor]. Was function SVGGElement() { [native code] }.
FAIL SVGDefsElement.toString() should be [object SVGDefsElementConstructor]. Was function SVGDefsElement() { [native code] }.
FAIL SVGDescElement.toString() should be [object SVGDescElementConstructor]. Was function SVGDescElement() { [native code] }.
FAIL SVGTitleElement.toString() should be [object SVGTitleElementConstructor]. Was function SVGTitleElement() { [native code] }.
FAIL SVGSymbolElement.toString() should be [object SVGSymbolElementConstructor]. Was function SVGSymbolElement() { [native code] }.
FAIL SVGUseElement.toString() should be [object SVGUseElementConstructor]. Was function SVGUseElement() { [native code] }.
FAIL SVGElementInstance.toString() should be [object SVGElementInstanceConstructor]. Was function SVGElementInstance() { [native code] }.
FAIL SVGElementInstanceList.toString() should be [object SVGElementInstanceListConstructor]. Was function SVGElementInstanceList() { [native code] }.
FAIL SVGImageElement.toString() should be [object SVGImageElementConstructor]. Was function SVGImageElement() { [native code] }.
FAIL SVGSwitchElement.toString() should be [object SVGSwitchElementConstructor]. Was function SVGSwitchElement() { [native code] }.
FAIL SVGStyleElement.toString() should be [object SVGStyleElementConstructor]. Was function SVGStyleElement() { [native code] }.
FAIL SVGPoint.toString() should be [object SVGPointConstructor]. Was function SVGPoint() { [native code] }.
FAIL SVGPointList.toString() should be [object SVGPointListConstructor]. Was function SVGPointList() { [native code] }.
FAIL SVGMatrix.toString() should be [object SVGMatrixConstructor]. Was function SVGMatrix() { [native code] }.
FAIL SVGTransform.toString() should be [object SVGTransformConstructor]. Was function SVGTransform() { [native code] }.
FAIL SVGTransformList.toString() should be [object SVGTransformListConstructor]. Was function SVGTransformList() { [native code] }.
FAIL SVGAnimatedTransformList.toString() should be [object SVGAnimatedTransformListConstructor]. Was function SVGAnimatedTransformList() { [native code] }.
FAIL SVGPreserveAspectRatio.toString() should be [object SVGPreserveAspectRatioConstructor]. Was function SVGPreserveAspectRatio() { [native code] }.
FAIL SVGAnimatedPreserveAspectRatio.toString() should be [object SVGAnimatedPreserveAspectRatioConstructor]. Was function SVGAnimatedPreserveAspectRatio() { [native code] }.
FAIL SVGPathSeg.toString() should be [object SVGPathSegConstructor]. Was function SVGPathSeg() { [native code] }.
FAIL SVGPathSegClosePath.toString() should be [object SVGPathSegClosePathConstructor]. Was function SVGPathSegClosePath() { [native code] }.
FAIL SVGPathSegMovetoAbs.toString() should be [object SVGPathSegMovetoAbsConstructor]. Was function SVGPathSegMovetoAbs() { [native code] }.
FAIL SVGPathSegMovetoRel.toString() should be [object SVGPathSegMovetoRelConstructor]. Was function SVGPathSegMovetoRel() { [native code] }.
FAIL SVGPathSegLinetoAbs.toString() should be [object SVGPathSegLinetoAbsConstructor]. Was function SVGPathSegLinetoAbs() { [native code] }.
FAIL SVGPathSegLinetoRel.toString() should be [object SVGPathSegLinetoRelConstructor]. Was function SVGPathSegLinetoRel() { [native code] }.
FAIL SVGPathSegCurvetoCubicAbs.toString() should be [object SVGPathSegCurvetoCubicAbsConstructor]. Was function SVGPathSegCurvetoCubicAbs() { [native code] }.
FAIL SVGPathSegCurvetoCubicRel.toString() should be [object SVGPathSegCurvetoCubicRelConstructor]. Was function SVGPathSegCurvetoCubicRel() { [native code] }.
FAIL SVGPathSegCurvetoQuadraticAbs.toString() should be [object SVGPathSegCurvetoQuadraticAbsConstructor]. Was function SVGPathSegCurvetoQuadraticAbs() { [native code] }.
FAIL SVGPathSegCurvetoQuadraticRel.toString() should be [object SVGPathSegCurvetoQuadraticRelConstructor]. Was function SVGPathSegCurvetoQuadraticRel() { [native code] }.
FAIL SVGPathSegArcAbs.toString() should be [object SVGPathSegArcAbsConstructor]. Was function SVGPathSegArcAbs() { [native code] }.
FAIL SVGPathSegArcRel.toString() should be [object SVGPathSegArcRelConstructor]. Was function SVGPathSegArcRel() { [native code] }.
FAIL SVGPathSegLinetoHorizontalAbs.toString() should be [object SVGPathSegLinetoHorizontalAbsConstructor]. Was function SVGPathSegLinetoHorizontalAbs() { [native code] }.
FAIL SVGPathSegLinetoHorizontalRel.toString() should be [object SVGPathSegLinetoHorizontalRelConstructor]. Was function SVGPathSegLinetoHorizontalRel() { [native code] }.
FAIL SVGPathSegLinetoVerticalAbs.toString() should be [object SVGPathSegLinetoVerticalAbsConstructor]. Was function SVGPathSegLinetoVerticalAbs() { [native code] }.
FAIL SVGPathSegLinetoVerticalRel.toString() should be [object SVGPathSegLinetoVerticalRelConstructor]. Was function SVGPathSegLinetoVerticalRel() { [native code] }.
FAIL SVGPathSegCurvetoCubicSmoothAbs.toString() should be [object SVGPathSegCurvetoCubicSmoothAbsConstructor]. Was function SVGPathSegCurvetoCubicSmoothAbs() { [native code] }.
FAIL SVGPathSegCurvetoCubicSmoothRel.toString() should be [object SVGPathSegCurvetoCubicSmoothRelConstructor]. Was function SVGPathSegCurvetoCubicSmoothRel() { [native code] }.
FAIL SVGPathSegCurvetoQuadraticSmoothAbs.toString() should be [object SVGPathSegCurvetoQuadraticSmoothAbsConstructor]. Was function SVGPathSegCurvetoQuadraticSmoothAbs() { [native code] }.
FAIL SVGPathSegCurvetoQuadraticSmoothRel.toString() should be [object SVGPathSegCurvetoQuadraticSmoothRelConstructor]. Was function SVGPathSegCurvetoQuadraticSmoothRel() { [native code] }.
FAIL SVGPathSegList.toString() should be [object SVGPathSegListConstructor]. Was function SVGPathSegList() { [native code] }.
FAIL SVGAnimatedPathData.toString() should be [object SVGAnimatedPathDataConstructor]. Threw exception ReferenceError: SVGAnimatedPathData is not defined
FAIL SVGPathElement.toString() should be [object SVGPathElementConstructor]. Was function SVGPathElement() { [native code] }.
FAIL SVGRectElement.toString() should be [object SVGRectElementConstructor]. Was function SVGRectElement() { [native code] }.
FAIL SVGCircleElement.toString() should be [object SVGCircleElementConstructor]. Was function SVGCircleElement() { [native code] }.
FAIL SVGEllipseElement.toString() should be [object SVGEllipseElementConstructor]. Was function SVGEllipseElement() { [native code] }.
FAIL SVGLineElement.toString() should be [object SVGLineElementConstructor]. Was function SVGLineElement() { [native code] }.
FAIL SVGAnimatedPoints.toString() should be [object SVGAnimatedPointsConstructor]. Threw exception ReferenceError: SVGAnimatedPoints is not defined
FAIL SVGPolylineElement.toString() should be [object SVGPolylineElementConstructor]. Was function SVGPolylineElement() { [native code] }.
FAIL SVGPolygonElement.toString() should be [object SVGPolygonElementConstructor]. Was function SVGPolygonElement() { [native code] }.
FAIL SVGTextContentElement.toString() should be [object SVGTextContentElementConstructor]. Was function SVGTextContentElement() { [native code] }.
FAIL SVGTextPositioningElement.toString() should be [object SVGTextPositioningElementConstructor]. Was function SVGTextPositioningElement() { [native code] }.
FAIL SVGTextElement.toString() should be [object SVGTextElementConstructor]. Was function SVGTextElement() { [native code] }.
FAIL SVGTSpanElement.toString() should be [object SVGTSpanElementConstructor]. Was function SVGTSpanElement() { [native code] }.
FAIL SVGTRefElement.toString() should be [object SVGTRefElementConstructor]. Was function SVGTRefElement() { [native code] }.
FAIL SVGTextPathElement.toString() should be [object SVGTextPathElementConstructor]. Was function SVGTextPathElement() { [native code] }.
FAIL SVGAltGlyphElement.toString() should be [object SVGAltGlyphElementConstructor]. Was function SVGAltGlyphElement() { [native code] }.
FAIL SVGAltGlyphDefElement.toString() should be [object SVGAltGlyphDefElementConstructor]. Was function SVGAltGlyphDefElement() { [native code] }.
FAIL SVGAltGlyphItemElement.toString() should be [object SVGAltGlyphItemElementConstructor]. Was function SVGAltGlyphItemElement() { [native code] }.
FAIL SVGGlyphRefElement.toString() should be [object SVGGlyphRefElementConstructor]. Was function SVGGlyphRefElement() { [native code] }.
FAIL SVGPaint.toString() should be [object SVGPaintConstructor]. Was function SVGPaint() { [native code] }.
FAIL SVGMarkerElement.toString() should be [object SVGMarkerElementConstructor]. Was function SVGMarkerElement() { [native code] }.
FAIL SVGColorProfileElement.toString() should be [object SVGColorProfileElementConstructor]. Threw exception ReferenceError: SVGColorProfileElement is not defined
FAIL SVGColorProfileRule.toString() should be [object SVGColorProfileRuleConstructor]. Threw exception ReferenceError: SVGColorProfileRule is not defined
FAIL SVGGradientElement.toString() should be [object SVGGradientElementConstructor]. Was function SVGGradientElement() { [native code] }.
FAIL SVGLinearGradientElement.toString() should be [object SVGLinearGradientElementConstructor]. Was function SVGLinearGradientElement() { [native code] }.
FAIL SVGRadialGradientElement.toString() should be [object SVGRadialGradientElementConstructor]. Was function SVGRadialGradientElement() { [native code] }.
FAIL SVGStopElement.toString() should be [object SVGStopElementConstructor]. Was function SVGStopElement() { [native code] }.
FAIL SVGPatternElement.toString() should be [object SVGPatternElementConstructor]. Was function SVGPatternElement() { [native code] }.
FAIL SVGClipPathElement.toString() should be [object SVGClipPathElementConstructor]. Was function SVGClipPathElement() { [native code] }.
FAIL SVGMaskElement.toString() should be [object SVGMaskElementConstructor]. Was function SVGMaskElement() { [native code] }.
FAIL SVGFilterElement.toString() should be [object SVGFilterElementConstructor]. Was function SVGFilterElement() { [native code] }.
FAIL SVGFilterPrimitiveStandardAttributes.toString() should be [object SVGFilterPrimitiveStandardAttributesConstructor]. Threw exception ReferenceError: SVGFilterPrimitiveStandardAttributes is not defined
FAIL SVGFEBlendElement.toString() should be [object SVGFEBlendElementConstructor]. Was function SVGFEBlendElement() { [native code] }.
FAIL SVGFEColorMatrixElement.toString() should be [object SVGFEColorMatrixElementConstructor]. Was function SVGFEColorMatrixElement() { [native code] }.
FAIL SVGFEComponentTransferElement.toString() should be [object SVGFEComponentTransferElementConstructor]. Was function SVGFEComponentTransferElement() { [native code] }.
FAIL SVGComponentTransferFunctionElement.toString() should be [object SVGComponentTransferFunctionElementConstructor]. Was function SVGComponentTransferFunctionElement() { [native code] }.
FAIL SVGFEFuncRElement.toString() should be [object SVGFEFuncRElementConstructor]. Was function SVGFEFuncRElement() { [native code] }.
FAIL SVGFEFuncGElement.toString() should be [object SVGFEFuncGElementConstructor]. Was function SVGFEFuncGElement() { [native code] }.
FAIL SVGFEFuncBElement.toString() should be [object SVGFEFuncBElementConstructor]. Was function SVGFEFuncBElement() { [native code] }.
FAIL SVGFEFuncAElement.toString() should be [object SVGFEFuncAElementConstructor]. Was function SVGFEFuncAElement() { [native code] }.
FAIL SVGFECompositeElement.toString() should be [object SVGFECompositeElementConstructor]. Was function SVGFECompositeElement() { [native code] }.
FAIL SVGFEConvolveMatrixElement.toString() should be [object SVGFEConvolveMatrixElementConstructor]. Was function SVGFEConvolveMatrixElement() { [native code] }.
FAIL SVGFEDiffuseLightingElement.toString() should be [object SVGFEDiffuseLightingElementConstructor]. Was function SVGFEDiffuseLightingElement() { [native code] }.
FAIL SVGFEDistantLightElement.toString() should be [object SVGFEDistantLightElementConstructor]. Was function SVGFEDistantLightElement() { [native code] }.
FAIL SVGFEPointLightElement.toString() should be [object SVGFEPointLightElementConstructor]. Was function SVGFEPointLightElement() { [native code] }.
FAIL SVGFESpotLightElement.toString() should be [object SVGFESpotLightElementConstructor]. Was function SVGFESpotLightElement() { [native code] }.
FAIL SVGFEDisplacementMapElement.toString() should be [object SVGFEDisplacementMapElementConstructor]. Was function SVGFEDisplacementMapElement() { [native code] }.
FAIL SVGFEFloodElement.toString() should be [object SVGFEFloodElementConstructor]. Was function SVGFEFloodElement() { [native code] }.
FAIL SVGFEGaussianBlurElement.toString() should be [object SVGFEGaussianBlurElementConstructor]. Was function SVGFEGaussianBlurElement() { [native code] }.
FAIL SVGFEImageElement.toString() should be [object SVGFEImageElementConstructor]. Was function SVGFEImageElement() { [native code] }.
FAIL SVGFEMergeElement.toString() should be [object SVGFEMergeElementConstructor]. Was function SVGFEMergeElement() { [native code] }.
FAIL SVGFEMergeNodeElement.toString() should be [object SVGFEMergeNodeElementConstructor]. Was function SVGFEMergeNodeElement() { [native code] }.
FAIL SVGFEMorphologyElement.toString() should be [object SVGFEMorphologyElementConstructor]. Was function SVGFEMorphologyElement() { [native code] }.
FAIL SVGFEOffsetElement.toString() should be [object SVGFEOffsetElementConstructor]. Was function SVGFEOffsetElement() { [native code] }.
FAIL SVGFESpecularLightingElement.toString() should be [object SVGFESpecularLightingElementConstructor]. Was function SVGFESpecularLightingElement() { [native code] }.
FAIL SVGFETileElement.toString() should be [object SVGFETileElementConstructor]. Was function SVGFETileElement() { [native code] }.
FAIL SVGFETurbulenceElement.toString() should be [object SVGFETurbulenceElementConstructor]. Was function SVGFETurbulenceElement() { [native code] }.
FAIL SVGCursorElement.toString() should be [object SVGCursorElementConstructor]. Was function SVGCursorElement() { [native code] }.
FAIL SVGAElement.toString() should be [object SVGAElementConstructor]. Was function SVGAElement() { [native code] }.
FAIL SVGViewElement.toString() should be [object SVGViewElementConstructor]. Was function SVGViewElement() { [native code] }.
FAIL SVGScriptElement.toString() should be [object SVGScriptElementConstructor]. Was function SVGScriptElement() { [native code] }.
FAIL SVGEvent.toString() should be [object SVGEventConstructor]. Threw exception ReferenceError: SVGEvent is not defined
FAIL SVGZoomEvent.toString() should be [object SVGZoomEventConstructor]. Was function SVGZoomEvent() { [native code] }.
FAIL SVGAnimationElement.toString() should be [object SVGAnimationElementConstructor]. Threw exception ReferenceError: SVGAnimationElement is not defined
FAIL SVGAnimateElement.toString() should be [object SVGAnimateElementConstructor]. Was function SVGAnimateElement() { [native code] }.
FAIL SVGSetElement.toString() should be [object SVGSetElementConstructor]. Was function SVGSetElement() { [native code] }.
FAIL SVGAnimateMotionElement.toString() should be [object SVGAnimateMotionElementConstructor]. Was function SVGAnimateMotionElement() { [native code] }.
FAIL SVGMPathElement.toString() should be [object SVGMPathElementConstructor]. Was function SVGMPathElement() { [native code] }.
FAIL SVGAnimateColorElement.toString() should be [object SVGAnimateColorElementConstructor]. Was function SVGAnimateColorElement() { [native code] }.
FAIL SVGAnimateTransformElement.toString() should be [object SVGAnimateTransformElementConstructor]. Was function SVGAnimateTransformElement() { [native code] }.
FAIL SVGFontElement.toString() should be [object SVGFontElementConstructor]. Was function SVGFontElement() { [native code] }.
FAIL SVGGlyphElement.toString() should be [object SVGGlyphElementConstructor]. Was function SVGGlyphElement() { [native code] }.
FAIL SVGMissingGlyphElement.toString() should be [object SVGMissingGlyphElementConstructor]. Was function SVGMissingGlyphElement() { [native code] }.
FAIL SVGHKernElement.toString() should be [object SVGHKernElementConstructor]. Was function SVGHKernElement() { [native code] }.
FAIL SVGVKernElement.toString() should be [object SVGVKernElementConstructor]. Was function SVGVKernElement() { [native code] }.
FAIL SVGFontFaceElement.toString() should be [object SVGFontFaceElementConstructor]. Was function SVGFontFaceElement() { [native code] }.
FAIL SVGFontFaceSrcElement.toString() should be [object SVGFontFaceSrcElementConstructor]. Was function SVGFontFaceSrcElement() { [native code] }.
FAIL SVGFontFaceUriElement.toString() should be [object SVGFontFaceUriElementConstructor]. Was function SVGFontFaceUriElement() { [native code] }.
FAIL SVGFontFaceFormatElement.toString() should be [object SVGFontFaceFormatElementConstructor]. Was function SVGFontFaceFormatElement() { [native code] }.
FAIL SVGFontFaceNameElement.toString() should be [object SVGFontFaceNameElementConstructor]. Was function SVGFontFaceNameElement() { [native code] }.
FAIL SVGDefinitionSrcElement.toString() should be [object SVGDefinitionSrcElementConstructor]. Threw exception ReferenceError: SVGDefinitionSrcElement is not defined
FAIL SVGMetadataElement.toString() should be [object SVGMetadataElementConstructor]. Was function SVGMetadataElement() { [native code] }.
FAIL SVGForeignObjectElement.toString() should be [object SVGForeignObjectElementConstructor]. Was function SVGForeignObjectElement() { [native code] }.
PASS successfullyParsed is true

TEST COMPLETE
